+++ src/containers/AddMessage.js
@@ -2,11 +2,17 @@ import { connect } from 'react-redux';
 import AddMessageComponent from '../components/AddMessage';
 import { addMessage } from '../actions';
 
+const mapDispatchToState = state => {
+  return ({
+    currentUser: state.currentUser
+  });
+};
+
 const mapDispatchToProps = dispatch => ({
-  dispatchMessage: (message, author) => dispatch(addMessage(message, author))
+  dispatchMessage: (message, uuid, author) => dispatch(addMessage(message, uuid, author))
 });
 
 export default connect(
-  null,
+  mapDispatchToState,
   mapDispatchToProps
 )(AddMessageComponent);